Update
Radiance is not just a direct increase (to 55%) from the base rate-up (50%) for the character banner, but is like a pity system.

The first rate-up attempt loss is the standard 50/50. Second has a 5% chance to trigger radiance when losing, therefore a 525/475 split. Third is a 75/25, and fourth is a 100% chance for winning the featured character.

Ditched the database approach while the file size appeared unsustainable.

Stopped centering the form.