Update index.html
Slight change for the notice when calculating.

The "very high" number for the calculator to take more than two seconds to finish could be claimed as being somewhere around 40 or more copies total from either characters or weapons, but it can't be so certainly declared, what with so many different devices around.

At 80 copies of characters and 5 copies of weapons, there's a case for it taking about 10 seconds.